requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.4
django-phone-field==1.8.1
django-recaptcha==3.0.0
django-recaptcha3==0.4.0
gunicorn==20.1.0
idna==3.3
lockfile==0.12.2
oauthlib==3.2.0
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0
